backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-socketio==5.10.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
firebase-admin==6.3.0
google-cloud-speech==2.23.0
google-cloud-translate==3.13.0
google-cloud-texttospeech==2.15.0
langdetect==1.0.9
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
